requirements.txt
python-dotenv==1.0.1
requests==2.31.0
spotipy==2.22.1


kafka-python==2.0.2
boto3==1.28.28
snowflake-connector-python==3.0.4

dbt-core==1.5.2
dbt-snowflake==1.5.2

apache-airflow==2.6.2

pyspark==3.4.1

